Remove old test code
There is some code that I had there after messing with some of the
topics on Uros' blog (https://popovicu.com/), but that became irrelevant
after working on some other projects (e.g. fbos) or on other examples.

This commit also brings about a major overhaul on the structure of the
project. Hopefully it's now at a more ready state.

Signed-off-by: Miquel Sabaté Solà <mikisabate@gmail.com>