supervisor: Add an example on the HSM extension
The HSM extension from SBI v2.0 allows a supervisor to manage the state
of harts on the system. Moreover, it also guarantees that only one hart
makes it into the booting stages of the supervisor, and it's up to the
supervisor to wake up the rest through the HSM extension.

I have added an example which showcases this extension in different
ways, and it finally shuts down the machine.

Signed-off-by: Miquel Sabaté Solà <mikisabate@gmail.com>